requirements.txt
selenium>=4.23.0
webdriver-manager>=4.0.2
pytest>=8.2.0
pytest-html>=4.1.1
python-dotenv>=1.0.1
